+++ models/chat/telegram.py
@@ -1,6 +1,7 @@
 from re import compile as re_compile
 from requests import get, exceptions, Timeout
 
+
 class Telegram():
     """Telegram client"""
     def __init__(self, token='', client_id=''):
@@ -17,12 +18,10 @@ class Telegram():
         if not p.match(client_id):
             raise Exception('Telegram client_id is invalid')
 
-        #print('Telegram configure with for client "' + client_id + '" with token "' + token + '"')
-
     def send(self, message='', parsemode='Markdown') -> str:
         """Send telegram message"""
         try:
-            escaped_message = message.translate(message.maketrans({"*":  r"\*"}))
+            escaped_message = message.translate(message.maketrans({"*": r"\*"}))
             payload = f'{self.api}{self._token}/sendMessage?chat_id={self._client_id}&parse_mode={parsemode}&text={escaped_message}'
             resp = get(payload)