+++ models/config/__init__.py
@@ -1,5 +1,5 @@
-from .binance_parser import parser as binanceConfigParser, parseMarket as binanceParseMarket
-from .coinbase_pro_parser import parser as coinbaseProConfigParser, parseMarket as coinbaseProParseMarket
-from .kucoin_parser import parser as kucoinConfigParser, parseMarket as kucoinParseMarket
-from .dummy_parser import parser as dummyConfigParser, parseMarket as dummyParseMarket
+from .binance_parser import parser as binanceConfigParser, parse_market as binanceParseMarket
+from .coinbase_pro_parser import parser as coinbaseProConfigParser, parse_market as coinbaseProParseMarket
+from .kucoin_parser import parser as kucoinConfigParser, parse_market as kucoinParseMarket
+from .dummy_parser import parser as dummyConfigParser, parse_market as dummyParseMarket
 from .logger_parser import parser as loggerConfigParser