+++ models/config/binance_parser.py
@@ -3,10 +3,10 @@ import json
 import os.path
 import re
 
-from .default_parser import isCurrencyValid, defaultConfigParse, merge_config_and_args
+from .default_parser import is_currency_valid, default_config_parse, merge_config_and_args
 
 
-def isMarketValid(market) -> bool:
+def is_market_valid(market) -> bool:
     if market is None:
         return False
 
@@ -18,130 +18,150 @@ def isMarketValid(market) -> bool:
         return True
     return False
 
-def parseMarket(market):
-    base_currency = 'BTC'
-    quote_currency = 'GBP'
 
-    if not isMarketValid(market):
-        raise ValueError(f'Binance market invalid: {market}')
+def parse_market(market):
+    base_currency = "BTC"
+    quote_currency = "GBP"
+
+    if not is_market_valid(market):
+        raise ValueError(f"Binance market invalid: {market}")
 
     quote_currencies = [
-        'BTC', 'BNB', 'ETH', 'USDT', 'TUSD', 'BUSD', 'DAX', 'NGN', 'RUB', 'TRY', 'EUR',
-        'GBP', 'ZAR', 'UAH', 'DAI', 'BIDR', 'AUD', 'USD', 'NGN', 'BRL', 'BVND', 'VAI'
+        "BTC",
+        "BNB",
+        "ETH",
+        "USDT",
+        "TUSD",
+        "BUSD",
+        "DAX",
+        "NGN",
+        "RUB",
+        "TRY",
+        "EUR",
+        "GBP",
+        "ZAR",
+        "UAH",
+        "DAI",
+        "BIDR",
+        "AUD",
+        "USD",
+        "NGN",
+        "BRL",
+        "BVND",
+        "VAI",
     ]
 
     for qc in quote_currencies:
-        if market.endswith( qc ):
-            base_currency = market.replace(qc, '')
+        if market.endswith(qc):
+            base_currency = market.replace(qc, "")
             quote_currency = qc
             break
 
     if len(market) != len(base_currency) + len(quote_currency):
-        raise ValueError('Binance market error.')
+        raise ValueError("Binance market error.")
 
     return market, base_currency, quote_currency
 
-def parser(app, binance_config, args={}):
-    #print('Binance Configuration parse')
 
+def parser(app, binance_config, args={}):
     if not app:
-        raise Exception('No app is passed')
+        raise Exception("No app is passed")
 
     if isinstance(binance_config, dict):
-        if 'api_key' in binance_config or 'api_secret' in binance_config:
-            print(f'>>> migrating api keys to binance.key <<<\n')
+        if "api_key" in binance_config or "api_secret" in binance_config:
+            print(">>> migrating api keys to binance.key <<<\n")
 
             # create 'binance.key'
-            fh = open('binance.key', 'w')
+            fh = open("binance.key", "w")
             fh.write(f"{binance_config['api_key']}\n{binance_config['api_secret']}")
             fh.close()
 
-            if os.path.isfile('config.json') and os.path.isfile('binance.key'):
-                binance_config['api_key_file'] = binance_config.pop('api_key')
-                binance_config['api_key_file'] = 'binance.key'
-                del binance_config['api_secret']
+            if os.path.isfile("config.json") and os.path.isfile("binance.key"):
+                binance_config["api_key_file"] = binance_config.pop("api_key")
+                binance_config["api_key_file"] = "binance.key"
+                del binance_config["api_secret"]
 
                 # read 'binance' element from config.json
-                fh = open('config.json', 'r')
+                fh = open("config.json", "r")
                 config_json = ast.literal_eval(fh.read())
-                config_json['binance'] = binance_config
+                config_json["binance"] = binance_config
                 fh.close()
 
                 # write new 'binance' element
-                fh = open('config.json', 'w')
+                fh = open("config.json", "w")
                 fh.write(json.dumps(config_json, indent=4))
                 fh.close()
             else:
-                print (f'migration failed (io error)\n')
+                print("migration failed (io error)\n")
 
         api_key_file = None
-        if 'api_key_file' in args and args['api_key_file'] is not None:
-            api_key_file = args['api_key_file']
-        elif 'api_key_file' in binance_config:
-            api_key_file = binance_config['api_key_file']
+        if "api_key_file" in args and args["api_key_file"] is not None:
+            api_key_file = args["api_key_file"]
+        elif "api_key_file" in binance_config:
+            api_key_file = binance_config["api_key_file"]
 
         if api_key_file is not None:
-            try :
-                with open( api_key_file, 'r') as f :
+            try:
+                with open(api_key_file, "r") as f:
                     key = f.readline().strip()
                     secret = f.readline().strip()
-                binance_config['api_key'] = key
-                binance_config['api_secret'] = secret
-            except :
+                binance_config["api_key"] = key
+                binance_config["api_secret"] = secret
+            except Exception:
                 raise RuntimeError(f"Unable to read {api_key_file}")
 
-        if 'api_key' in binance_config and 'api_secret' in binance_config and 'api_url' in binance_config:
+        if "api_key" in binance_config and "api_secret" in binance_config and "api_url" in binance_config:
             # validates the api key is syntactically correct
             p = re.compile(r"^[A-z0-9]{64,64}$")
-            if not p.match(binance_config['api_key']):
-                raise TypeError('Binance API key is invalid')
+            if not p.match(binance_config["api_key"]):
+                raise TypeError("Binance API key is invalid")
 
-            self.api_key = binance_config['api_key']
+            api_key = binance_config["api_key"]  # noqa: F841
 
             # validates the api secret is syntactically correct
             p = re.compile(r"^[A-z0-9]{64,64}$")
-            if not p.match(binance_config['api_secret']):
-                raise TypeError('Binance API secret is invalid')
+            if not p.match(binance_config["api_secret"]):
+                raise TypeError("Binance API secret is invalid")
 
-            self.api_secret = binance_config['api_secret']
+            api_secret = binance_config["api_secret"]  # noqa: F841
 
             valid_urls = [
-                'https://api.binance.com/',
-                'https://testnet.binance.vision/',
-                'https://api.binance.com',
-                'https://testnet.binance.vision',
-                'https://api.binance.us',
+                "https://api.binance.com/",
+                "https://testnet.binance.vision/",
+                "https://api.binance.com",
+                "https://testnet.binance.vision",
+                "https://api.binance.us",
             ]
 
             # validate Binance API
-            if binance_config['api_url'] not in valid_urls:
-                raise ValueError('Binance API URL is invalid')
+            if binance_config["api_url"] not in valid_urls:
+                raise ValueError("Binance API URL is invalid")
 
-            self.api_url = binance_config['api_url']
-            self.base_currency = 'BTC'
-            self.quote_currency = 'GBP'
+            api_url = binance_config["api_url"]  # noqa: F841
+            base_currency = "BTC"
+            quote_currency = "GBP"
     else:
         binance_config = {}
 
     config = merge_config_and_args(binance_config, args)
 
-    defaultConfigParse(app, config)
+    default_config_parse(app, config)
 
-    if 'base_currency' in config and config['base_currency'] is not None:
-        if not isCurrencyValid(config['base_currency']):
-            raise TypeError('Base currency is invalid.')
-        self.base_currency = config['base_currency']
+    if "base_currency" in config and config["base_currency"] is not None:
+        if not is_currency_valid(config["base_currency"]):
+            raise TypeError("Base currency is invalid.")
+        base_currency = config["base_currency"]
 
-    if 'quote_currency' in config and config['quote_currency'] is not None:
-        if not isCurrencyValid(config['quote_currency']):
-            raise TypeError('Quote currency is invalid.')
-        self.quote_currency = config['quote_currency']
+    if "quote_currency" in config and config["quote_currency"] is not None:
+        if not is_currency_valid(config["quote_currency"]):
+            raise TypeError("Quote currency is invalid.")
+        quote_currency = config["quote_currency"]
 
-    if 'market' in config and config['market'] is not None:
-        self.market, self.base_currency, self.quote_currency = parseMarket(config['market'])
+    if "market" in config and config["market"] is not None:
+        market, base_currency, quote_currency = parse_market(config["market"])
 
-    if self.base_currency != '' and self.quote_currency != '':
-        self.market = self.base_currency + self.quote_currency
+    if base_currency != "" and quote_currency != "":
+        market = base_currency + quote_currency  # noqa: F841
 
-    if 'use_sell_fee' in config:
-        self.use_sell_fee = config['use_sell_fee']
+    if "use_sell_fee" in config:
+        use_sell_fee = config["use_sell_fee"]  # noqa: F841