+++ models/config/kucoin_parser.py
@@ -4,23 +4,24 @@ import json
 import os.path
 import sys
 
-from .default_parser import isCurrencyValid, defaultConfigParse, merge_config_and_args
+from .default_parser import is_currency_valid, default_config_parse, merge_config_and_args
 from models.exchange.Granularity import Granularity
 
-def isMarketValid(market) -> bool:
+
+def is_market_valid(market) -> bool:
     p = re.compile(r"^[0-9A-Z]{1,20}\-[1-9A-Z]{2,5}$")
     return p.match(market) is not None
 
-def parseMarket(market):
-    if not isMarketValid(market):
+
+def parse_market(market):
+    if not is_market_valid(market):
         raise ValueError('Kucoin market invalid: ' + market)
 
     base_currency, quote_currency = market.split('-', 2)
     return market, base_currency, quote_currency
 
-def parser(app, kucoin_config, args={}):
-    #print('Kucoin Configuration parse')
 
+def parser(app, kucoin_config, args={}):
     if not app:
         raise Exception('No app is passed')
 
@@ -50,7 +51,7 @@ def parser(app, kucoin_config, args={}):
                 fh.write(json.dumps(config_json, indent=4))
                 fh.close()
             else:
-                print ('migration failed (io error)', "\n")
+                print('migration failed (io error)', "\n")
 
         api_key_file = None
         if 'api_key_file' in args and args['api_key_file'] is not None:
@@ -60,14 +61,14 @@ def parser(app, kucoin_config, args={}):
 
         if api_key_file is not None:
             try :
-                with open( api_key_file, 'r') as f :
+                with open(api_key_file, 'r') as f :
                     key = f.readline().strip()
                     secret = f.readline().strip()
                     password = f.readline().strip()
                 kucoin_config['api_key'] = key
                 kucoin_config['api_secret'] = secret
                 kucoin_config['api_passphrase'] = password
-            except :
+            except Exception:
                 raise RuntimeError(f"Unable to read {api_key_file}")
 
         if 'api_key' in kucoin_config and 'api_secret' in kucoin_config and \
@@ -77,21 +78,21 @@ def parser(app, kucoin_config, args={}):
             if not p.match(kucoin_config['api_key']):
                 raise TypeError('Kucoin API key is invalid')
 
-            self.api_key = kucoin_config['api_key']
+            api_key = kucoin_config['api_key']  # noqa: F841
 
             # validates the api secret is syntactically correct
             p = re.compile(r"^[A-z0-9-]{36,36}$")
             if not p.match(kucoin_config['api_secret']):
                 raise TypeError('Kucoin API secret is invalid')
 
-            self.api_secret = kucoin_config['api_secret']
+            api_secret = kucoin_config['api_secret']  # noqa: F841
 
             # validates the api passphrase is syntactically correct
             p = re.compile(r"^[A-z0-9#$%=@!{},`~&*()<>?.:;_|^/+\[\]]{8,32}$")
             if not p.match(kucoin_config['api_passphrase']):
                 raise TypeError('Kucoin API passphrase is invalid')
 
-            self.api_passphrase = kucoin_config['api_passphrase']
+            api_passphrase = kucoin_config['api_passphrase']  # noqa: F841
 
             valid_urls = [
                 'https://api.kucoin.com/',
@@ -104,34 +105,34 @@ def parser(app, kucoin_config, args={}):
             if kucoin_config['api_url'] not in valid_urls:
                 raise ValueError('Kucoin API URL is invalid')
 
-            self.api_url = kucoin_config['api_url']
-            self.base_currency = 'BTC'
-            self.quote_currency = 'GBP'
+            api_url = kucoin_config['api_url']  # noqa: F841
+            base_currency = 'BTC'
+            quote_currency = 'GBP'
     else:
         kucoin_config = {}
 
     config = merge_config_and_args(kucoin_config, args)
 
-    defaultConfigParse(app, config)
+    default_config_parse(app, config)
 
     if 'base_currency' in config and config['base_currency'] is not None:
-        if not isCurrencyValid(config['base_currency']):
+        if not is_currency_valid(config['base_currency']):
             raise TypeError('Base currency is invalid.')
-        self.base_currency = config['base_currency']
+        base_currency = config['base_currency']
 
     if 'quote_currency' in config and config['quote_currency'] is not None:
-        if not isCurrencyValid(config['quote_currency']):
+        if not is_currency_valid(config['quote_currency']):
             raise TypeError('Quote currency is invalid.')
-        self.quote_currency = config['quote_currency']
+        quote_currency = config['quote_currency']
 
     if 'market' in config and config['market'] is not None:
-        self.market, self.base_currency, self.quote_currency = parseMarket(config['market'])
+        market, base_currency, quote_currency = parse_market(config['market'])
 
-    if self.base_currency != '' and self.quote_currency != '':
-        self.market = self.base_currency + '-' + self.quote_currency
+    if base_currency != '' and quote_currency != '':
+        market = base_currency + '-' + quote_currency  # noqa: F841
 
     if 'granularity' in config and config['granularity'] is not None:
         if isinstance(config['granularity'], str) and config['granularity'].isnumeric() is True:
-            self.granularity = Granularity.convert_to_enum(int(config['granularity']))
+            granularity = Granularity.convert_to_enum(int(config['granularity']))
         elif isinstance(config['granularity'], int):
-            self.granularity = Granularity.convert_to_enum(config['granularity'])
+            granularity = Granularity.convert_to_enum(config['granularity'])  # noqa: F841