+++ models/config/logger_parser.py
@@ -1,54 +1,53 @@
 import re
 
-from .default_parser import isCurrencyValid, defaultConfigParse, merge_config_and_args
+from .default_parser import is_currency_valid, default_config_parse, merge_config_and_args
 from models.helper.LogHelper import Logger
 
-def parser(app, logger_config):
-    #print('Logger Configuration parser')
 
+def parser(app, logger_config):
     if not logger_config:
-        raise Exception('There is an error in your config dictionnary')
+        raise Exception("There is an error in your config dictionnary")
 
     if not app:
-        raise Exception('No app is passed')
+        raise Exception("No app is passed")
 
-    if 'filelog' in logger_config:
-        if isinstance(logger_config['filelog'], int):
-            if logger_config['filelog'] in [0, 1]:
-                self.filelog = logger_config['filelog']
+    if "filelog" in logger_config:
+        if isinstance(logger_config["filelog"], int):
+            if logger_config["filelog"] in [0, 1]:
+                app.filelog = logger_config["filelog"]
         else:
-            raise TypeError('filelog must be type of int')
+            raise TypeError("filelog must be type of int")
 
-    if self.filelog:
-        if 'logfile' in logger_config:
-            if isinstance(logger_config['logfile'], str):
-                if self.logfile == "pycryptobot.log":
-                    self.logfile = logger_config['logfile']
+    if app.filelog:
+        if "logfile" in logger_config:
+            if isinstance(logger_config["logfile"], str):
+                if app.logfile == "pycryptobot.log":
+                    app.logfile = logger_config["logfile"]
             else:
-                raise TypeError('logfile must be type of str')
+                raise TypeError("logfile must be type of str")
 
-        if 'fileloglevel' in logger_config:
-            if isinstance(logger_config['fileloglevel'], str):
-                if logger_config['fileloglevel'] in ('CRITICAL', 'ERROR', 'WARNING', 'INFO', 'DEBUG', 'NOTSET'):
-                    self.fileloglevel = logger_config['fileloglevel']
+        if "fileloglevel" in logger_config:
+            if isinstance(logger_config["fileloglevel"], str):
+                if logger_config["fileloglevel"] in ("CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG", "NOTSET"):
+                    app.fileloglevel = logger_config["fileloglevel"]
                 else:
                     raise TypeError('fileloglevel must be one of: "CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG", "NOTSET"')
             else:
-                raise TypeError('fileloglevel must be type of str')
+                raise TypeError("fileloglevel must be type of str")
 
-    if 'consolelog' in logger_config:
-        if isinstance(logger_config['consolelog'], int):
-            if logger_config['consolelog'] in [0, 1]:
-                self.consolelog = logger_config['consolelog']
+    if "consolelog" in logger_config:
+        if isinstance(logger_config["consolelog"], int):
+            if logger_config["consolelog"] in [0, 1]:
+                app.consolelog = logger_config["consolelog"]
         else:
-            raise TypeError('consolelog must be type of int')
+            raise TypeError("consolelog must be type of int")
 
-    if self.consolelog:
-        if 'consoleloglevel' in logger_config:
-            if isinstance(logger_config['consoleloglevel'], str):
-                if logger_config['consoleloglevel'] in ('CRITICAL', 'ERROR', 'WARNING', 'INFO', 'DEBUG', 'NOTSET'):
-                    self.consoleloglevel = logger_config['consoleloglevel']
+    if app.consolelog:
+        if "consoleloglevel" in logger_config:
+            if isinstance(logger_config["consoleloglevel"], str):
+                if logger_config["consoleloglevel"] in ("CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG", "NOTSET"):
+                    app.consoleloglevel = logger_config["consoleloglevel"]
                 else:
                     raise TypeError('consoleloglevel must be one of: "CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG", "NOTSET"')
             else:
-                raise TypeError('consoleloglevel must be type of str')
+                raise TypeError("consoleloglevel must be type of str")